refactored ReaderUpdateService business logic out to ReaderUpdateLogic
