+++ WordPress/src/main/java/org/wordpress/android/ui/reader/services/update/ServiceCompletionListener.java
@@ -1,0 +1,5 @@
+package org.wordpress.android.ui.reader.services.update;
+
+public interface ServiceCompletionListener {
+    void onCompleted();
+}
